+++ CHANGES.md
@@ -12,6 +12,7 @@ Release Notes.
 #### Plugins
 * Support [go-redis](https://github.com/redis/go-redis) v9 redis client framework.
 * Support collecting [Native HTTP](https://pkg.go.dev/net/http) URI parameter on server side.
+* Support [Mongo](https://github.com/mongodb/mongo-go-driver) database client framework.
 
 #### Documentation